pset9/project/test_project.py
# Import necessary modules for testing and functions from the 'project' module
import os
import pytest
from project import (
    caesar_encrypt,
    caesar_decrypt,
    sha256_encrypt,
    encrypt_with_private_key,
    decrypt_with_private_key,
    generate_rsa_key_pair,
    rsa_encrypt,
    decrypt_with_rsa_private_key,
    save_private_key_to_file,
    load_private_key_from_file,
    save_encrypted_text_to_file,
)

# Define test functions for different encryption and decryption methods
def test_caesar_encrypt_decrypt():
    text = "Hello, World!"
    rotation = 3
    encrypted_text = caesar_encrypt(text, rotation)
    decrypted_text = caesar_decrypt(encrypted_text, rotation)
    assert decrypted_text == text

def test_sha256_encrypt():
    text = "Hello, World!"
    encrypted_text = sha256_encrypt(text)
    # As SHA-256 is a one-way hash, we can't decrypt it, so just check for non-empty string
    assert encrypted_text

def test_encrypt_decrypt_with_private_key():
    private_key = b'01234567890123456789012345678901'  # Replace with your Fernet key
    text = "Hello, World!"
    encrypted_text = encrypt_with_private_key(text, private_key)
    decrypted_text = decrypt_with_private_key(encrypted_text, private_key)
    assert decrypted_text == text

def test_rsa_encrypt_decrypt():
    text = "Hello, World!"
    private_key, public_key = generate_rsa_key_pair()
    encrypted_text = rsa_encrypt(text, public_key)
    decrypted_text = decrypt_with_rsa_private_key(encrypted_text, private_key)
    assert decrypted_text == text

def test_save_load_private_key():
    private_key = b'01234567890123456789012345678901'  # Replace with your Fernet key
    save_private_key_to_file(private_key)
    loaded_private_key = load_private_key_from_file()
    assert loaded_private_key == private_key

def test_save_load_encrypted_text():
    private_key = b'01234567890123456789012345678901'  # Replace with your Fernet key
    text = "Hello, World!"
    encrypted_text = encrypt_with_private_key(text, private_key)
    save_encrypted_text_to_file(encrypted_text)
    with open("encrypted_message.txt", "rb") as file:
        loaded_encrypted_text = file.read()
    assert loaded_encrypted_text == encrypted_text

# Add more tests as needed

# Run the tests using pytest when the script is executed
if __name__ == "__main__":
    pytest.main()

"""
The code imports necessary modules for testing (pytest) and functions from the 'project' module.
Test functions are defined for different encryption and decryption methods, including Caesar cipher, SHA-256 hash, Fernet symmetric encryption, and RSA encryption.
The pytest.main() is called to execute the tests when the script is run. Each test function contains assertions to check if the implemented functions produce the expected results.
The script can be extended by adding more test functions as needed for additional functionalities or edge cases.
"""